Move readableBytes utility function to existing  file (review feedback)
